Renamed dbghelp.h to avoid possible confusion
