+++ src/vld.vcxproj.filters
@@ -89,7 +89,7 @@
     <ClInclude Include="vldallocator.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="dbghelp.h">
+    <ClInclude Include="vld_dbghelp.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\setup\version.h">
